requirements.txt
fastapi
uvicorn
motor
pymongo
PyGithub
pydantic
django-environ